### 2024-02-15 #6344 @jbytecode
@m-dadej - Hi, thank you for submitting your work to JOSS. I am the handling editor. Do you have any suggestion for reviewers? You can use the potential reviewers list provided at the top of this issue. I think people interested in statistics and Julia would be a good fit. If you have suggestions, please provide a list of potential reviewers here but do not use the `@` character to avoid receiving unwanted notifications. Thank you in advance.
