+++ paper/paper.bib
@@ -3,6 +3,7 @@
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
   volume={59},
+  DOI = {10.1137/141000671},
   number={1},
   pages={65--98},
   year={2017},
@@ -17,6 +18,7 @@
  author = {James D. Hamilton},
  journal = {Econometrica},
  number = {2},
+ DOI = {10.2307/1912559},
  pages = {357--384},
  publisher = {[Wiley, Econometric Society]},
  title = {A New Approach to the Economic Analysis of Nonstationary Time Series and the Business Cycle},
@@ -46,7 +48,7 @@ number = {4},
 pages = {944-962},
 year = {2014},
 issn = {0169-2070},
-doi = {https://doi.org/10.1016/j.ijforecast.2014.03.014},
+doi = {10.1016/j.ijforecast.2014.03.014},
 url = {https://www.sciencedirect.com/science/article/pii/S0169207014000612},
 author = {Patrick T. Brandt and John R. Freeman and Philip A. Schrodt},
 keywords = {Conflict dynamics, Bayesian, Time series, Density evaluation, Verification rank histogram, Scoring rules},
@@ -60,7 +62,7 @@ volume = {47},
 pages = {101374},
 year = {2023},
 issn = {2214-5818},
-doi = {https://doi.org/10.1016/j.ejrh.2023.101374},
+doi = {10.1016/j.ejrh.2023.101374},
 url = {https://www.sciencedirect.com/science/article/pii/S2214581823000617},
 author = {Huimin Wang and Songbai Song and Gengxi Zhang and Olusola O. Ayantoboc},
 keywords = {Weihe River Basin, Daily streamflow, Heteroscedasticity, Multi-regime, MS-GARCH}
@@ -73,7 +75,7 @@ volume = {6},
 pages = {100059},
 year = {2021},
 issn = {2590-0544},
-doi = {https://doi.org/10.1016/j.csfx.2021.100059},
+doi = {10.1016/j.csfx.2021.100059},
 url = {https://www.sciencedirect.com/science/article/pii/S259005442100004X},
 author = {Yegnanew A. Shiferaw},
 keywords = {Case fatality rate, Coronavirus, COVID-19, Global, Markov-switching autoregressive approach}
@@ -87,7 +89,7 @@ number = {4},
 pages = {101081},
 year = {2020},
 issn = {1751-1577},
-doi = {https://doi.org/10.1016/j.joi.2020.101081},
+doi = {10.1016/j.joi.2020.101081},
 url = {https://www.sciencedirect.com/science/article/pii/S1751157720301851},
 author = {Fernando Delbianco and Andrés Fioriti and Allan Hernandez-Chanto and Fernando Tohmé},
 keywords = {Markov process, Regime switching, Citations, Growth, Stability},
@@ -117,6 +119,7 @@ title = {What's for dynr: A package for linear and nonlinear
   dynamic modeling in R},
 author = {Lu Ou and Michael D. Hunter and Sy-Miin Chow},
 journal = {The R Journal},
+DOI = {10.32614/rj-2019-012},
 year = {2019},
 volume = {11},
 pages = {1-20}
@@ -125,6 +128,7 @@ pages = {1-20}
 @inproceedings{statsmodels,
   title={statsmodels: Econometric and statistical modeling with python},
   author={Seabold, Skipper and Perktold, Josef},
+  DOI = {10.25080/majora-92bf1922-011},
   booktitle={9th Python in Science Conference},
   year={2010},
 }
@@ -134,9 +138,11 @@ pages = {1-20}
[…]
