### 2024-03-01 #6344 @jbytecode
👋👋👋 Dear @harishsiravuri and @ManuelStapper 👋👋👋

**[MarSwitching.jl: A Julia package for Markov switching dynamic models](https://github.com/openjournals/joss-reviews/issues/6344)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6344).

Thank you in advance!
