### 2024-03-04 #6441 @editorialbot
[REVIEW]: MarSwitching.jl: A Julia package for Markov switching dynamic models
**Submitting author:** @m-dadej (<a href="http://orcid.org/0000-0002-1791-7611">Mateusz Dadej</a>)
**Repository:** https://github.com/m-dadej/MarSwitching.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.3
