### 2024-03-18 7e0af86
correct issue: https://github.com/m-dadej/MarSwitching.jl/issues/8
