+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"man/examples/example/#Regime-switching-Phillips-curve","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"","category":"section"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"One of the most popular macroeconomic relationships is the trade-off between inflation and unemployment. The so-called Phillips curve is discussed in both introductory macroeconomics courses and at the meetings of the most influential central banks. The curve introduced a stylized fact that the inflation falls during recessions and rises during booms.","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"However, many policymakers and academic economists have argued that the historical relationship has changed over time. The 'flattening' of the Phillips curve poses a challenge for policymakers, as it can imply that countercyclical policy may not be effective in steering inflation toward the established central bank's target.","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"To investigate the time-varying nature of the Phillips curve, we will estimate a Markov switching model.","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"First we would need a dataset with quarterly inflation and unemployment. We will use the data from the Federal Reserve Bank of St. Louis (FRED) database. The data is available in the repository of the package. It's already transformed into log differences.","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"using MarSwitching\nusing DataFrames\nusing CSV\nusing Plots\nusing Random\nusing Dates\n\ndf = CSV.read(\"my_assets/philips.csv\", DataFrame, missingstring = \"NA\")\n\nmodel_df = dropmissing(select(df, [:cpi, :unemp,  :infexp, :s_shock]))  ","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"Let's see how the relationship looks like in the raw data:","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"using Plots\n\nplot_df = filter(x -> x.cpi .> -2, model_df) # remove outliers\n\nphil_plot = plot(plot_df.unemp, plot_df.cpi,\n                seriestype=:scatter, legend = :none,\n                title = \"Phillips curve\",\n                xlabel = \"Unemployment rate gap\", ylabel = \"CPI change\")","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"(Image: Plot)","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"Overall, the relationship is far from being clear. The slope of plotted data is just slightly negative. At least for these ","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"x = [ones(size(model_df)[1]) model_df.unemp]\n(x'x)^(-1)*(x'model_df.cpi)","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips curve","title":"Regime switching Phillips curve","text":"2-element Vector{Float64}:\n  0.926584897227734\n -0.0024658207247481023","category":"page"},{"location":"man/examples/example/","page":"Regime switching Phillips cu
[…]
