### 2024-03-23 #6441 @y1my1
Overall, this is a good mature package for achieving what the author intends to do. I could tell the author has put a lot of effort into programming the software and writing the documentation. I have some minor issues with the following

### Functionality

I was trying to verify whether the package can replicate the results of a basic regression model I found [here](https://www.statsmodels.org/dev/examples/notebooks/generated/markov_regression.html) (data could be found [here](http://www.stata-press.com/data/r14/usmacro.dta)). However, the results I got from this package seem to be quite different from what I can get from statsmodel (Python) and Stata (statsmodel and Stata gave almost the same results). See attached graph below

I tried to compare the performance of the corresponding methods in the package and statsmodel (Python). I generated a simulated data of 10000 rows, following [this example](https://github.com/m-dadej/MarSwitching.jl/tree/main?tab=readme-ov-file#example). The results don't seem to be that different, at least on my MacBook (see below). However, I may not be doing it the right way; it would be good if the author could document how he did the performance test.

### Documentation

There are no community guidelines, but this should be an easy fix.
