### 2024-04-02 #6441 @m-dadej
I have not seen any package that do identical task. There is a package to model hidden Markov models (https://github.com/gdalle/HiddenMarkovModels.jl) which may allow similar estimation. However as you can see in its [example](https://gdalle.github.io/HiddenMarkovModels.jl/dev/examples/controlled/#Learning), it requires additional programming from user in order to estimate a rather simple Markov switching model. Do you have any other package in mind that I missed?
