### 2024-04-02 #6441 @jbytecode
@m-dadej  - Yes, you can refer to the package, emphasizing that it can address the same types of problems with additional effort, while highlighting that your package is capable of efficiently handling these tasks in a smoother manner rather than a sharp one like the current one. Thank you in advance.
