### 2024-04-12 571ac85
hiddenmarkovmodels.jl reference
