+++ paper/paper.bib
@@ -143,6 +143,18 @@ pages = {1-20}
   url = {https://github.com/msperlin/MS\_Regress-Matlab}
 }
 
+@article{Dalle2024, 
+  doi = {10.21105/joss.06436}, 
+  url = {https://doi.org/10.21105/joss.06436}, 
+  year = {2024},
+  publisher = {The Open Journal}, 
+  volume = {9}, 
+  number = {96}, 
+  pages = {6436}, 
+  author = {Guillaume Dalle}, 
+  title = {HiddenMarkovModels.jl: generic, fast and reliable state space modeling},
+  journal = {Journal of Open Source Software}
+}
 
 @article{filardo94,
  ISSN = {07350015},
