### 2024-05-05 #6441 @mkitti
For the examples, an environment including a Project.toml and Manifest.toml should be included with the repository. This would allow user to reproduce the examples with the exact versions of the packages used.

Some of the examples use data included in the package as follows:

```julia
df = CSV.read("my_assets/philips.csv", DataFrame, missingstring = "NA")

df = CSV.read("my_assets/df_spx.csv", DataFrame)
```

I might make those relative to the path of the package that the user installed so that no additional work needs to be done for the user to try the examples. They would otherwise have to clone the repository, but this is not necessary since the files are already installed.

```julia
philips_csv_path = joinpath(
    dirname(pathof(MarSwitching)),
    "../docs/src/man/examples/my_assets/philips.csv"
)
df = CSV.read(philips_csv_path, DataFrame, missingstring = "NA")

df_spx_csv_path = joinpath(
    dirname(pathof(MarSwitching)),
    "../docs/src/man/examples/my_assets/df_spx.csv"
)
df = CSV.read(df_spx_csv_path, DataFrame)
```
