### 2024-05-05 #6441 @mkitti
The reference and limitations of HiddenMarkovModels.jl would be greatly appreciated. Users may be trying to decide between the two packages and a comparison of this package that would be very useful.
