### 2024-05-05 #6441 @mkitti
I see Community Guidelines in the README.md of the Github repository, but not in the documentation itself.

https://github.com/m-dadej/MarSwitching.jl/blob/main/README.md#contributing

Please include this information in https://m-dadej.github.io/MarSwitching.jl/
