### 2024-05-05 #6441 @mkitti
Overall, MarSwitching.jl is an important contribution to the field, allowing Markov switching dynamic models to be simulated and analyzed in the Julia Programming Langauge. Comparisons are provided to packages in other langauges. Comprehensive examples are provided. The documentation and paper are well written and organized.

Additional comparisons and contrasts to existing packages in the Julia ecosystem would be appreciated as others have noted in this review. I also suggested a few minor changes to make it easier to run the examples by referencing data files relative to the installed package.

If the concerns raised above in my own review and those are the other reviewer are addressed, I would recommend the paper for publication.

I apologize for the delay in reviewing the submission.
