### 2024-05-05 #6441 @mkitti
Minor correction

> MarSwitching.jl is a package for estimating Markov switching dynamic models (also called regime switching) for Julia. This is a class of models with time-varying coefficients depending on an unobservable state/regime that follows [a] Markov process. The package provides tools for estimation, inference and simulation of the models.

The "a" should be inserted.
