### 2024-05-05 #6441 @y1my1
> Thank you for your comments [y1my1](https://github.com/y1my1) and sorry for delayed response, I was busy with my Phd.
>
> Regarding the functionality comment. The way you defined a model with `MSModel` function won't give you the same one as in statsmodels. Here is a correct command:
>
> `model = MSModel(df[2:end, 2], 2, exog_switching_vars = df[begin:end-1, 2], switching_var = false)`
>
> `exog_vars` is for non-switching exogenous variables and as the equation in the example suggests the variance of the model is not regime-dependent (thus I specified `switching_var = false`). A default value is a regime-dependent variance. I am very much open to suggestion on how to make the function more intuitive to use :)
>
> That being said, a current version of the package still gives some of the estimates a bit off than in statsmodels:
>
> ![image](https://private-user-images.githubusercontent.com/48299815/318115908-8ae458da-8c57-4dbb-82d9-41730a06ca20.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTQ5MzYyMTIsIm5iZiI6MTcxNDkzNTkxMiwicGF0aCI6Ii80ODI5OTgxNS8zMTgxMTU5MDgtOGFlNDU4ZGEtOGM1Ny00ZGJiLTgyZDktNDE3MzBhMDZjYTIwLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA1MDUlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNTA1VDE5MDUxMlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWQzYmYzOTU1ZmU4ZTdhZWYxNDcyNjQxNTIwMTEyNmU4ZDUwMTI4ZDgyZTczNTE1MTMxNjkyODUyYzMxZWJmZmEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.qTByERBHd8-9FoQpLRUcLhaF69YD2TaI7Kdu60xy_lg)
>
> The one intercept now is way different. I think some time ago I compared it but disregarded the issue as the rest of the examples matches my package. Plus, the Matlab package `MS_Regress` also had different estimates. But anyway, I checked my code and the issue was solved in the most recent commit ([m-dadej/MarSwitching.jl@6fa582d](https://github.com/m-dadej/MarSwitching.jl/commit/6fa582df3fedd16ae64e5ffeaa5ba194f362226a)). (tl;dr Before the commit, I was adding a small number to state probability to solve an underflow error that have been happening some time ago).
>
> If you install the github version of the package you can now replicate the results from statsmodels.
>
> Regarding the performance comparison. The code for benchmark is in [benchmark folder](https://github.com/m-dadej/MarSwitching.jl/tree/main/benchmark). The readme example can give different results as benchmark (in terms of relative performance), because the model is different (less complex). In the benchmark code I estimate 3 regime, 1 non-switching and 1 switching exogenous variable model.
>
> That being said, I compare the readme example and the relative performance between statsmodel is smaller. MarSwitching is ~2x faster. But I am unable to replicate your results, even though the code on your screen shoot seems to be fine.
>
> About community guidelines, it's a good idea. I will write them soon.

The performance difference we saw may be because of the different environments we are using. I updated Julia to the newest version (1.10.3) and was able to see MarSwitching runs 3 times faster compared with statsmodel. I would recommend adding a note of what environment is used. In addition, I am not sure I see people use "," as equivalent to "." anywhere, e.g., "6,7", which is hard to read, at least to me.
