### 2024-05-07 41fcbff
examples path and environment
