### 2024-05-07 7d38fba
HiddenMarkovModels.jl comparison in SoN
