### 2024-05-08 #6441 @m-dadej
Hey @jbytecode I added changes according to the reveiwer'ssuggestions.

@mkitti project.toml and manifest.toml are now in the examples folder and the scripts now have the path generating function you sugested. The contribution guidelines are also in docs now.

About the HiddenMarkovModels.jl comparison. The last paragraph in statement of need now reads:

>Despite the popularity of the method, `MarSwitching.jl` is, at the moment, the only package that allows for effortless estimation of Markov switching models with Julia programming language. At the same time, it is implemented purely in this language. For more general modeling with hidden Markov models, Julia users may find `HiddenMarkovModels.jl` [Dalle2024] package useful as well. `HiddenMarkovModels.jl` offers more generic approach to programming hidden Markov models, albeit requiring user-side development of certain estimation algorithms for Markov switching models, as well as model inference functions.

I did not elaborate further as the paper is already pretty long. Of course I am open to any suggestions.

@y1my1 I added the information regarding the environment. The paragraph below the benchmark table now reads:

>Software versions: MarSwitching.jl v0.2.2, Statsmodels v0.14.1, MSwM v1.5, MS_Regress v1.11. The programming languages versions were: Julia v1.10.1, Python v3.12.2, R v4.2.1 and MATLAB vR2024a. Calculations were run on: Windows 11 x64 Intel(R) Core(TM) i7-9850H CPU @ 2.60GHz, 2592 Mhz, 6 Core(s), 12 Logical Processor(s).

The commas are removed as well.
