### 2024-05-23 #6441 @mkitti
The mention and comparison to HiddenMarkovModels.jl is very welcome, and I appreciate the author adding this paragraph. Ultimately, I think it will be very helpful for readers in deciding between packages

> Despite the popularity of the method, `MarSwitching.jl` is, at the moment, the only package that allows for effortless estimation of Markov switching models with Julia programming language.

The "effortless" seems hyperbolic here. I would appreciate if the author could replace effortless by something more specific, for example:

> `MarSwitching.jl` is, at the moment, the only package that allows for estimation of Markov switching models with Julia programming language by specifying a minimal set of regime switching parameters.
