### 2024-05-26 dfb3e08
substitute "effortless" according to suggestion
