### 2024-05-26 #6441 @m-dadej
Hey @mkitti, @jbytecode

I replaced the sentence accordingly. I think the one you suggested is ok. Thank you
