### 2024-06-04 #6441 @jbytecode
@m-dadej - We have a compiling issue of the pdf. The error message indicates

```
Warning:  Could not convert image /tmp/tex2pdf.-74ed9d6bad7c1236/regime_probs.svg: check that rsvg-convert is in path.
  rsvg-convert: createProcess: posix_spawnp: does not exist (No such file or directory)
```

so I think it is about rendering the svg image. Is it possible to convert it to png or use the png version of the image?
