### 2024-06-05 #6441 @m-dadej
I changed to .png. It should be ok now
