### 2024-06-06 #6441 @jbytecode
@m-dadej - I've just sent a pull request. This PR includes some minor changes and fixes in both the bibtex and manuscript files. Please review the changes. If you are agreed with them, please merge. Thank you in advance.
