### 2024-06-07 #6441 @jbytecode
@m-dadej - The sentence

`Where the former is estimated using the data up to time $t$ and the latter using the whole dataset. The functions to get these probabilities are `filtered_probs()` and `smoothed_probs()` respectively.`

doesn't sound clear to me, and I am not really sure I understand the point there, so could you please refactor this sentence and get it more clear?

After that, we need to create a version tag, e.g. `vx.y.z`, in the target repository. Please use this version tag to create a new Zenodo archive. Please provide me the versiyon tag and the zenodo link here.

In Zenodo archive, the list of authors, the title of the paper, and the ORCIDs should match exactly with those in the submission.

Thank you in advance.
