### 2024-06-10 332c216
smoothed probability sentance
