### 2024-06-10 #6441 @jbytecode
@m-dadej - The license of the software is MIT but in Zenodo archive, it appears `
Creative Commons Attribution 4.0 International`. Could you please fix?
