+++ docs/src/man/examples/example_spx.md
@@ -24,7 +24,7 @@ df_spx_csv_path = joinpath(
 df = CSV.read(df_spx_csv_path, DataFrame)
 ```
 
-The model we are going to estimate might be challenging for the optimizing algorithm, thus we can help it by standardizing the variables. This is especially important for the models with time-varying transition probability matrix.
+The model we are going to estimate might be challenging for the optimizing algorithm, thus we can help it by standardizing the variables. When the covariates are unscaled, the corresponding parameters may inappropriately dominate the gradient. This is especially important for the models with time-varying transition probability matrix. 
 
 ```jldoctest spx
 σ_spx = std(df.spx)
