### 2024-06-12 #6441 @jbytecode
@m-dadej - Sorry, I've just realized that the missing DOI is actively directed to the package, so we can fix it. Please merge the pull request. Then we can go forward
