### 2024-06-12 c37bf66
replace dfrac with frac in manuscript
