### 2024-06-12 #6441 @jbytecode
@m-dadej - Sorry for the inconvenience. The final renderer doesn't accept `dfrac` or `mfrac` but using `frac` is okay. The PR fixes (possibly) this issue. Could you please merge it?
