### 2024-06-12 #6441 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
