### 2024-06-12 #6441 @crvernon
:wave: @m-dadej - This one is looking very good.  Just a few minor things in the paper:

- [ ] LINE 6: "package allows users of Julia programming language" should be "package allows users of the Julia programming language" where "the" is added in.
- [ ] LINES 10-12:  Consider changing "Because of non-trivial likelihood function and the amount of model parameters, Julia is a perfect language to implement this class of models due to its performance." to "Given the non-trivial nature of the likelihood function and the amount of model parameters, Julia is an ideal language for implementing this class of models due to its computational performance." for better readability.
- [ ] LINES 29-30:  "MATLAB users may also estimate these models with MS_Regress (Perlin, 2012) package." should be "MATLAB users may also estimate these models with the MS_Regress (Perlin, 2012) package." where "the" is added in.
- [ ] LINE 34: "with Julia" should be "with the Julia"
- [ ] LINE 37: "find HiddenMarkovModels.jl" should be "find the HiddenMarkovModels.jl"
- [ ] LINE 38: "offers more" should be "offers a more"
- [ ] LINE 47: "of dependent" should be "of the dependent"
- [ ] LINE 59:  are you missing a prime here after delta?  "where 𝛿𝑖,𝑗 is"
- [ ] LINES 69-70: "to either exog_vars or exog_switching_vars argument," should be "to either the exog_vars or the exog_switching_vars argument,"
- [ ] LINE 70: "pass exogenous" should be "pass an exogenous"
- [ ] LINE 70: "for time-varying" should be "for the time-varying"
- [ ] LINE 92: "the model requires also a forecast" should be "the model also requires a forecast"

Let me know once you have made these changes and we will proceed.  Thanks!
