### 2024-06-13 456e782
linear algebra typos
