### 2024-06-14 #6441 @jbytecode
@crvernon - The equation

```
\begin{equation*}
p_{i,j,t} = \dfrac{\exp(\mathbf{Z}_t \delta_{i,j})}{\textstyle \sum_{j=1} \exp(\mathbf{Z}_t \delta_{i,j})}
\end{equation*}
```

includes `dfrac` which is problematic, it should be `frac` to pass our system. It is interesting that I have recently sent a PR to fix these issue, however, the file seems got back to its older form.

@m-dadej - Did you revert the changes?
